Refactor SortCriteria to it's own file
